requirements.txt
colored==2.2.4
grpcio==1.56.0
grpcio-tools==1.56.0
tabulate==0.9.0
